feat: improvement of socialpages components
