+++ src/helpers/constant.ts
@@ -7,4 +7,8 @@ export const providers = [
   "Dealer Rater",
   "Facebook"
 ];
+export const selectProvider = [...providers.filter((item) => item !== "Cars")];
+
+export const filterProvider = ["All Social Media", ...selectProvider];
+
 export const ratings = [5, 4, 3, 2, 1];
